Extract goto test helper
