+++ crates/red_knot_ide/src/goto.rs
@@ -249,27 +249,18 @@ pub(crate) fn find_goto_target(parsed: &ParsedModule, offset: TextSize) -> Optio
 
 #[cfg(test)]
 mod tests {
-    use std::fmt::Write;
-
-    use crate::db::tests::TestDb;
+    use crate::tests::{cursor_test, CursorTest, IntoDiagnostic};
     use crate::{goto_type_definition, NavigationTarget};
     use insta::assert_snapshot;
-    use insta::internals::SettingsBindDropGuard;
-    use red_knot_python_semantic::{
-        Program, ProgramSettings, PythonPath, PythonPlatform, SearchPathSettings,
-    };
     use ruff_db::diagnostic::{
-        Annotation, Diagnostic, DiagnosticFormat, DiagnosticId, DisplayDiagnosticConfig, LintName,
-        Severity, Span, SubDiagnostic,
+        Annotation, Diagnostic, DiagnosticId, LintName, Severity, Span, SubDiagnostic,
     };
-    use ruff_db::files::{system_path_to_file, File, FileRange};
-    use ruff_db::system::{DbWithWritableSystem, SystemPath, SystemPathBuf};
-    use ruff_python_ast::PythonVersion;
-    use ruff_text_size::{Ranged, TextSize};
+    use ruff_db::files::FileRange;
+    use ruff_text_size::Ranged;
 
     #[test]
     fn goto_type_of_expression_with_class_type() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             class Test: ...
 
@@ -299,7 +290,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_function_type() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def foo(a, b): ...
 
@@ -331,7 +322,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_union_type() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
 
             def foo(a, b): ...
@@ -388,7 +379,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_module() {
-        let mut test = goto_test(
+        let mut test = cursor_test(
             r#"
             import lib
 
@@ -418,7 +409,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_literal_type() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             a: str = "test"
 
@@ -449,7 +440,7 @@ mod tests {
     }
     #[test]
     fn goto_type_of_expression_with_literal_node() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             a: str = "te<CURSOR>st"
             "#,
@@ -477,7 +468,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_type_var_type() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             type Alias[T: int = bool] = list[T<CURSOR>]
             "#,
@@ -501,7 +492,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_type_param_spec() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             type Alias[**P = [int, str]] = Callable[P<CURSOR>, int]
             "#,
@@ -515,7 +506,7 @@ mod tests {
 
     #[test]
     fn goto_type_of_expression_with_type_var_tuple() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             type Alias[*Ts = ()] = tuple[*Ts<CURSOR>]
             "#,
@@ -529,7 +520,7 @@ mod tests {
 
     #[test]
     fn goto_type_on_keyword_argument() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def test(a: str): ...
 
@@ -561,7 +552,7 @@ mod tests {
 
     #[test]
     fn goto_type_on_incorrectly_typed_keyword_argument() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def test(a: str): ...
 
@@ -596,7 +587,7 @@ mod tests {
 
     #[test]
     fn goto_type_on_kwargs() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def f(name: str): ...
 
@@ -630,7 +621,7 @@ f(**kwargs<CURSOR>)
 
     #[test]
     fn goto_type_of_expression_with_builtin() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def foo(a: str):
                 a<CURSOR>
@@ -661,7 +652,7 @@ f(**kwargs<CURSOR>)
 
     #[test]
     fn goto_type_definition_cursor_between_object_and_attribute() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             class X:
                 def foo(a, b): ...
@@ -693,7 +684,7 @@ f(**kwargs<CURSOR>)
 
     #[test]
     fn goto_between_call_arguments() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def foo(a, b): ...
 
@@ -723,7 +714,7 @@ f(**kwargs<CURSOR>)
 
     #[test]
     fn goto_type_narrowing() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def foo(a: str | None, b):
                 if a is not None:
@@ -755,7 +746,7 @@ f(**kwargs<CURSOR>)
 
     #[test]
     fn goto_type_none() {
-        let test = goto_test(
+        let test = cursor_test(
             r#"
             def foo(a: str | None, b):
                 a<CURSOR>
@@ -800,65 +791,7 @@ f(**kwargs<CURSOR>)
         ");
     }
 
-    fn goto_test(source: &str) -> GotoTest {
-        let mut db = TestDb::new();
-        let cursor_offset = source.find("<CURSOR>").expect(
-            "`source`` should contain a `<CURSOR>` marker, indicating the position of the cursor.",
-        );
-
-        let mut content = source[..cursor_offset].to_string();
-        content.push_str(&source[cursor_offset + "<CURSOR>".len()..]);
-
-        db.write_file("main.py", &content)
-            .expect("write to memory file system to be successful");
-
-        let file = system_path_to_file(&db, "main.py").expect("newly written file to existing");
-
-        Program::from_settings(
-            &db,
-            ProgramSettings {
-                python_version: PythonVersion::latest(),
-                python_platform: PythonPlatform::default(),
-                search_paths: SearchPathSettings {
-                    extra_paths: vec![],
-                    src_roots: vec![SystemPathBuf::from("/")],
-                    custom_typeshed: None,
-                    python_path: PythonPath::KnownSitePackages(vec![]),
-                },
-            },
-        )
-        .expect("Default settings to be valid");
-
-        let mut insta_settings = insta::Settings::clone_current();
-        insta_settings.add_filter(r#"\\(\w\w|\s|\.|")"#, "/$1");
-
-        let insta_settings_guard = insta_settings.bind_to_scope();
-
-        GotoTest {
-            db,
-            cursor_offset: TextSize::try_from(cursor_offset)
-                .expect("source to be smaller than 4GB"),
-            file,
-            _insta_settings_guard: insta_settings_guard,
-        }
-    }
-
-    struct GotoTest {
-        db: TestDb,
-        cursor_offset: TextSize,
-        file: File,
-        _insta_settings_guard: SettingsBindDropGuard,
-    }
-
-    impl GotoTest {
-        fn write_file(
-            &mut self,
-            path: impl AsRef<SystemPath>,
-            content: &str,
-        ) -> std::io::Result<()> {
-            self.db.write_file(path, content)
-        }
-
+    impl CursorTest {
         fn goto_type_definition(&self) -> String {
             let Some(targets) = goto_type_definition(&self.db, self.file, self.cursor_offset)
             else {
@@ -869,19 +802,12 @@ f(**kwargs<CURSOR>)
                 return "No type definitions found".to_string();
             }
 
-            let mut buf = String::new();
-
             let source = targets.range;
-
-            let config = DisplayDiagnosticConfig::default()
-                .color(false)
-                .format(DiagnosticFormat::Full);
-            for target in &*targets {
-                let diag = GotoTypeDefinitionDiagnostic::new(source, target).into_diagnostic();
-                write!(buf, "{}", diag.display(&self.db, &config)).unwrap();
-            }
-
-            buf
+            self.render_diagnostics(
+                targets
+                    .into_iter()
+                    .map(|target| GotoTypeDefinitionDiagnostic::new(source, &target)),
+            )
         }
     }
 
@@ -897,7 +823,9 @@ f(**kwargs<CURSOR>)
                 target: FileRange::new(target.file(), target.focus_range()),
             }
         }
+    }
 
+    impl IntoDiagnostic for GotoTypeDefinitionDiagnostic {
         fn into_diagnostic(self) -> Diagnostic {
             let mut source = SubDiagnostic::new(Severity::Info, "Source");
             source.annotate(Annotation::primary(
